Update to 1.1.4 	modified:   CMakeLists.txt 	modified:   LICENSE 	modified:   README.md 	modified:   gs_camera/__init__.py 	renamed:    gs_camera/core.py -> gs_camera/gs_camera.py 	modified:   launch/camera.launch 	modified:   launch/photo.launch 	modified:   package.xml 	modified:   setup.py
