+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package format="2">
   <name>gs_camera</name>
-  <version>1.1.3</version>
+  <version>1.1.4</version>
   <description>The gs_camera package</description>
 
   <maintainer email="ilya@eadsoft.com">Ilya Danilenko</maintainer>
